Adds new vertex attribute : ATTRIBUTE_MATERIAL_INDICES
updates internal shaders to use a new variable i_mtl containing the indices to use in the sampled array texture.

updates asset shaders to propage the new vevrtex attribute and use it.

updates render example to set the new attribute on the mesh vertices, with hardcoded indices [0,1,2,3] to account for the array textures including exactly 4 textures each.